+++ opportunities-service/src/main/java/com/ironhack/opportunitiesservice/service/interfaces/IOpportunityService.java
@@ -58,6 +58,6 @@ public interface IOpportunityService {
     int getMinQuantity();
     double getMedian(String data);
 
-    String findOpportunityCountBySalesRep(int salesRepId);
+    String findOpportunityCountBySalesRep(Long salesRepId);
     String findOpportunityByStatusCountBySalesRep(Status status);
 }
